Is wristband?
A wristband can have 4 patterns:
horizontal: each item in a row is identical.
vertical: each item in a column is identical.
diagonal left: each item is identical to the one on its upper left or bottom right.
diagonal right: each item is identical to the one on its upper right or bottom left.

This function that returns true if the section can be correctly classified into one of the 4 types, and false otherwise.